Create main.ml
Main file to parse a Roman numeral as a string.
Main evaluation function returns the answer as a 'result' type.
result has two variants - Ok(v) and Err(e). In this case v is of type int and e of type string.
v is the value when a valid Roman numeral is entered and evaluated.
e is a string containing information as to why the evaluation was unsuccessful -  this is either due to an unrecognised character or the Roman numeral input being an empty string to which no value has been assigned, instead of 0.

TODO:
i) refactor eval_num_list function - appears to contain some duplicated code and seems like the logic could be simplified.

ii) Create a simple REPL for Roman numeral valuation.